refactor: remove unused variants, consolidate appbar icon buttons
