+++ dev-client/src/screens/DeleteAccountScreen/DeleteAccountScreen.tsx
@@ -18,11 +18,11 @@
 import {useTranslation} from 'react-i18next';
 import {ScrollView} from 'react-native-gesture-handler';
 
+import {ScreenBackButton} from 'terraso-mobile-client/components/buttons/icons/appBar/ScreenBackButton';
 import {ScreenContentSection} from 'terraso-mobile-client/components/content/ScreenContentSection';
 import {Column} from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {useUserDeletionRequests} from 'terraso-mobile-client/hooks/userDeletionRequest.ts';
 import {AppBar} from 'terraso-mobile-client/navigation/components/AppBar';
-import {ScreenBackButton} from 'terraso-mobile-client/navigation/components/ScreenBackButton';
 import {DeleteAccountConfirmContent} from 'terraso-mobile-client/screens/DeleteAccountScreen/components/DeleteAccountConfirmContent';
 import {DeleteAccountConfirmForm} from 'terraso-mobile-client/screens/DeleteAccountScreen/components/DeleteAccountConfirmForm';
 import {DeleteAccountPendingContent} from 'terraso-mobile-client/screens/DeleteAccountScreen/components/DeleteAccountPendingContent';
